:recycle: break out press and award components
into separate files
